improve pia install script

+++ topics/pia/install.sh
@@ -1,3 +1,5 @@
-wget https://www.privateinternetaccess.com/installer/pia-nm.sh
-sudo bash ./pia-nm.sh
-rm -f pia-nm.sh
+if ! ls /etc/NetworkManager/system-connections | grep -q PIA; then
+  wget https://www.privateinternetaccess.com/installer/pia-nm.sh
+  sudo bash ./pia-nm.sh
+  rm -f pia-nm.sh
+fi
